requirements.txt
asgiref==3.4.1
boto3==1.20.35
botocore==1.23.35
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-braces==1.15.0
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-countdowntimer-model==0.0.7
django-easymode==0.14.5
django-environ==0.4.5
django-filter==21.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeltranslation==0.17.3
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-payments-paybox==0.2.3
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.1
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
environs==9.5.0
python-decouple